Feature core gameplay

+++ textAnimator.h
@@ -3,4 +3,4 @@
 #include <iostream>
 #include <string>
 
-void print_animated(std::string text, float seconds);
+void print_animated(std::string text, double seconds);
